// core/job/providers/enum/construction-property.enum.ts
export enum EConstructionProp {
  BURNT_CLAY_BRICKS = 'BURNT_CLAY_BRICKS',
  CONCRETE_BLOCKS = 'CONCRETE_BLOCKS',
  UNBURNT_CLAY_BLOCKS = 'UNBURNT_CLAY_BLOCKS',
  MUD_AND_WATTLE = 'MUD_AND_WATTLE',
  BURNT_BRICKS_UNBRICK_CLAY_BLOCKS = 'BURNT_BRICKS_UNBRICK_CLAY_BLOCKS',
  RULIBA_BRICKS = 'RULIBA_BRICKS'
}