// core/job/providers/enum/building-types.enum.ts
export enum EBuildingType {
  RESIDENCIAL = 'residencial',
  COMMERCIAL = 'commercial',
  INDUSTRIAL = 'industrial',
  RECREATIONAL_FACILITIES = 'recreational_facilities',
  SPECIAL_PROPERTIES = 'special_properties',
  AGRICULTURAL_BUILDING = 'agricultural_building'
}

export enum ELandType {
  VACANT_LAND = 'vacant_land',
  LAND_IMPROVEMENT = 'land_improvement',
  FOREST = 'forest',
  CULTIVABLE_LAND = 'cultivable_land',
  OTHER_NATURAL_RESOURCES = 'other_natural_resources'
}

export enum EResidenceCategory {
  APARTMENT = 'apartment',
  SINGLE_FAMILY_HOUSE = 'single_family_house',
  DUPLEX = 'duplex',
  OFFICES = 'offices',
  MALLS = 'malls',
  RETAIL_SHOPS_KIOSK = 'retail_shops_kiosk',
  HOTEL = 'hotel',
  WAREHOUSES = 'warehouses',
  GAS_PETROL_STATION = 'gas_petrol_station',
  STADIUMS = 'stadiums',
  AIRPORT = 'airport'
}

export enum EBuildingCat {
  BUILDING_CATEGORY = 'building_category'
}