// core/job/providers/enum/valuation-property.enum.ts
export enum EValuationProperty {
  STORED = 'stored',
  FLOOR = 'floor',
  BATHROOMS = 'bathrooms',
  BEDROOMS = 'bedrooms',
  KITCHEN = 'kitchen',
  OFFICE = 'office',
  DINING_ROOM = 'dining_room',
  LIVING_ROOM = 'living_room',
  LIFT = 'lift',
  SWIMMING_POOL = 'swimming_pool',
  GARDEN = 'garden',
  PAVED_AREA = 'paved_area',
  FENCE = 'fence',

  // new on duplex
  SHARED_WALL = 'shared_wall',
  SHARED_SWIMMING_POOL = 'shared_swimming_pool',
  SHARED_STAIRCASE = 'shared_staircase',

  // new on offices [commercial]
  OFFICE_SPACE = 'office_space',
  BOARDROOMS = 'boardrooms',
  HALLS = 'halls',
  TOILETS_BATHROOMS = 'toilets_bathrooms',
  AC = 'ac',
  PARKING_UNDERGROUND_OUTSIDE = 'parking_underground_outside',
  STAIRCASE = 'staircase',
  STORIES = 'stories',
  ARCHIVES_STORES = 'archives_stores',

  // new on malls
  STOREY = 'storey',
  TOILETES = 'toiletes',

  // new retail shop/kiosk
  SHOPS = 'shops',
  STORAGE_AREA = 'storage_area',

  // new  on hotel
  RESTAURANTS = 'restaurants',
  BARS = 'bars',
  SAUNA_STEAM = 'sauna_steam',
  GYM = 'gym',
  PLAYGROUND = 'playground',
  PARKING = 'parking',

  // new warehouses
  STORAGE_SPACE_SIZE = 'storage_space_size',

  // indutrial
  PRODUCTION_SPACE = 'production_space',
  RAW_MATERIAL_STORE = 'raw_material_store',
  WAREHOUSE = 'warehouse',
  CONTROL_ROOM = 'control_room',
  COLD_ROOM = 'cold_room',
  ADMINISTRATION_BLOCK = 'administration_block',
  DORMITORIES = 'dormitories',
  RESTAURANTS_CANTINE = 'restaurants_cantine',
  CLINIC_FIRSTAID = 'clinic_firstaid',

  // new on recreational facilities
  THEATERS = 'theaters',
  WATERFALLS = 'waterfalls',
  CANTINE = 'cantine',

  // new on special properties [Gas/petrol station]
  NUMBER_OF_PUMPS = 'number_of_pumps',
  SIZE_OF_STORAGE_SPACE_TANKS = 'size_of_storage_space_TANKS',
  TARMAC_SIZE = 'tarmac_size',
  ATTACHED_SHOPS = 'attached_shops',
  ATTACHED_GARAGE = 'attached_garage',
  OTHER_TANKS_FOR_OILS = 'other_tanks_for_oils',

  // new on special properties [Stadiums (specify on each item)]
  SITTING_SPACE = 'sitting_space',
  HOTELS = 'hotels',
  SAUNA_STEAM_MASSAGE = 'sauna_steam_massage',
  CHANGING_ROOMS = 'changing_rooms',

  // new on special properties [Airport]
  OFFICE_BLOCKS = 'office_blocks',

  // [LAND building type]
  // vacant land
  SIZE = 'size_area',

  // forest
  TREES_TYPE = 'trees_type',
  SIZE_OF_TREES = 'size_of_trees',
  DISTANCE = 'distance',

  // cultivable land
  TYPES_CROPS = 'types_crops',
  TYPE_OF_SOIL = 'type_of_soil',

  BUILT_UP_AREA = 'built_up_area',
  GROSS_FLOOR_AREA = 'gross_floor_area'

}